+++ tests/DetectorTest.php
@@ -57,7 +57,7 @@ class DetectorTest extends TestCase
                 ],
                 [
                     'line' => 50,
-                    'value' => -2,
+                    'value' => -1,
                 ],
             ],
             $fileReport->getEntries()
@@ -259,6 +259,24 @@ class DetectorTest extends TestCase
         $this->assertEmpty($fileReport->getEntries());
     }
 
+    public function testDetectReadingNumber(): void
+    {
+        $option = $this->createOption();
+        $option->setExtensions([new ArrayExtension]);
+        $option->setIncludeNumericStrings(true);
+        $detector = $this->createDetector($option);
+
+        $fileReport = $detector->detect(FileReportTest::getTestFile('test_1'));
+
+        $this->assertContains(
+            [
+                'line' => 64,
+                'value' => 1234,
+            ],
+            $fileReport->getEntries()
+        );
+    }
+
     public function testAllowArrayMappingWithArrayExtension(): void
     {
         $option = $this->createOption();
@@ -297,6 +315,60 @@ class DetectorTest extends TestCase
             [
                 'line' => 30,
                 'value' => 13,
+            ],
+            $fileReport->getEntries()
+        );
+    }
+
+
+    public function testDefaultIgnoreFunctions(): void
+    {
+        $option = $this->createOption();
+        $option->setExtensions([new ArrayExtension()]);
+        $option->setIncludeNumericStrings(true);
+        $detector = $this->createDetector($option);
+
+        $fileReport = $detector->detect(FileReportTest::getTestFile('test_1'));
+
+        $results = $fileReport->getEntries();
+
+        $this->assertNotContains(
+            [
+                'line' => 56,
+                'value' => 13,
+            ],
+            $results
+        );
+
+        $this->assertNotContains(
+            [
+                'line' => 57,
+                'value' => 3.14,
+            ],
+            $results
+        );
+
+        $this->assertNotContains(
+            [
+                'line' => 58,
+                'value' => 10,
+            ],
+            $results
+        );
+    }
+
+    public function testCheckForMagicArrayConstants(): void
+    {
+        $option = $this->createOption();
+        $option->setExtensions([new ArrayExtension()]);
+        $detector = $this->createDetector($option);
+
+        $fileReport = $detector->detect(FileReportTest::getTestFile('test_3'));
+
+        $this->assertContains(
+            [
+                'line' => 4,
+                'value' => 2,
             ],
             $fileReport->getEntries()
         );